+++ src/effect/index.ts
@@ -8,34 +8,15 @@
 // TODO: implement radial blur
 // TODO: implement zoom blur
 
-import Base from './base'
-import Brightness from './brightness'
-import Channels from './channels'
-import ChromaKey from './chroma-key'
-import Contrast from './contrast'
-import EllipticalMask from './elliptical-mask'
-import { GaussianBlur, GaussianBlurHorizontal, GaussianBlurVertical } from './gaussian-blur'
-import Grayscale from './grayscale'
-import Pixelate from './pixelate'
-import { Shader, TextureOptions, UniformOptions } from './shader'
-import Stack from './stack'
-import Transform from './transform'
-
-export {
-  Base,
-  Brightness,
-  Channels,
-  ChromaKey,
-  Contrast,
-  EllipticalMask,
-  GaussianBlur,
-  GaussianBlurHorizontal,
-  GaussianBlurVertical,
-  Grayscale,
-  Pixelate,
-  Shader,
-  TextureOptions,
-  UniformOptions,
-  Stack,
-  Transform
-}
+export * from './base'
+export * from './brightness'
+export * from './channels'
+export * from './chroma-key'
+export * from './contrast'
+export * from './elliptical-mask'
+export * from './gaussian-blur'
+export * from './grayscale'
+export * from './pixelate'
+export * from './shader'
+export * from './stack'
+export * from './transform'